+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add `a-tab-show` event to `a-tab` element, which is fired when the tab is shown and deprecate the `a-tab-select` event.
 - Add `a-tab-hide` event to `a-tab` element, which is fired when the tab is hidden.
 - Add `selectTabById()` method to `a-tab-group` element, which selects the tab with the given id.
+- Remove experimental view transitions on panels via `panel-transition` attribute.
 - Generate types declarations for the components.
 - Update dev dependencies.